requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
qdrant-client>=1.7.0
sentence-transformers>=2.3.0
rank-bm25>=0.2.2
langgraph
langchain
langchain-community
arize-phoenix
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
openinference-instrumentation-openai
datasets
langchain-openai
streamlit
requests
python-multipart
reportlab
mlx-lm
huggingface_hub
pypdf
